add continue to build.pl

+++ openwrt_cc_v3.0/build.pl
@@ -18,7 +18,6 @@ my @target = ();
 
 sub get_all_conffiles {
     my $tmp_dir = shift @_;
-    printf("tmp_dir=$tmp_dir\n");
     opendir(TEMP_DIR, $tmp_dir) || die "open $tmp_dir fail...$!";
     my @tmp_files = readdir TEMP_DIR;
     closedir(TEMP_DIR);
@@ -37,23 +36,25 @@ Version: $V
 
 Options:
   --target testapp(,extcc3.0...)   target various comma separated types
+  -c, --continue             continue build
   -h, --help, --version      display this help and exit
 
 for example:
   $P --target testapp
-  $P --target testapp,extcc3.0
+  $P --target testapp,extcc3.0,u01,inter_connd
 EOM
 
         exit($exitcode);
 }
 
 my $help = 0;
-printf("args=$#ARGV\n");
+my $continue = 0;
 if ($#ARGV < 0) {
     help(1);
 }
 GetOptions(
         'target=s'      => \@target,
+	'continue|c'      => \$continue,
         'help'        => \$help,
         'version'       => \$help
 ) or help(1);
@@ -65,13 +66,11 @@ help(0) if ($help);
 
 my $count = 0;
 
-printf("hello $root_dir\n $confs_dir\n");
-print("@target\n");
-
 foreach my $word (@target) {
         next if ($word =~ m/extcc/);
         next if ($word =~ m/testapp/);
         next if ($word =~ m/u01/);
+	next if ($word =~ m/inter_connd/);
 	printf("$word is not the right target!\n");
         exit(1);
 }
@@ -80,15 +79,23 @@ my @confs_files=get_all_conffiles($confs_dir);
 for my $filename (@confs_files) {
     next if ($filename eq '.' or $filename eq '..');
     next if (!($filename =~ m/^.+\.config$/));
-#    next if ($filename eq 'sd5117x.config');
+    next if ($filename eq 'hi5663.config');
 #    $count++;
 #    next if ($count < 25);
-#    next if ($filename eq 'zx279127.config');
-#    next if ($filename eq 'zx279128.config');
+    next if ($filename eq 'zx279127.config');
+    next if ($filename eq 'zx279128.config');
+    next if ($filename eq 'sd5117x.config');
+    next if ($filename eq 'zx279131.config');
     printf("compile $filename ...\n");
     system("rm -f .config") == 0 or die "[$filename] rm -f .config failed:$?";
     system("ln -sf confs/$filename .config") == 0 or die "[$filename] ln -sf confs/$filename .config failed:$?";
     foreach my $word (@target) {
+	if ($continue eq 1) {
+            my $temp = $filename =~ s/^(.*).config$/$1/r;
+            my $tmp_dir = "$root_dir/bin/$temp/packages/base";
+            my @file_list = glob "$tmp_dir/$word*$temp.ipk";
+            next if (@file_list);
+	}
         printf("make defconfig $filename $word\n");
         system("make defconfig") == 0 or die "[$filename:$word] make defconfig failed:$?";
         if ($word eq "testapp") {
@@ -100,13 +107,12 @@ for my $filename (@confs_files) {
         } elsif ($word eq "u01") {
             printf("$filename:$word\n");
             system("sed -i 's/# CONFIG_PACKAGE_u01v3 is not set/CONFIG_PACKAGE_u01v3=y/g' .config") == 0 or die "[$filename:$word] modify .config failed:$?";
+        } elsif ($word eq "inter_connd") {
+            printf("$filename:$word\n");
+            system("sed -i 's/# CONFIG_PACKAGE_inter_conndv3 is not set/CONFIG_PACKAGE_inter_conndv3=y/g' .config") == 0 or die "[$filename:$word] modify .config failed:$?";
         }
         system("make prepare") == 0 or die "[$filename:$word] make prepare failed:$?";
         system("make package/upointech/$word/clean") == 0 or die "[$filename:$word] make package/upointech/$word/clean failed:$?";
         system("make package/upointech/$word/compile") == 0 or die "[$filename:$word] make package/upointech/$word/compile failed:$?";;
     }
-#    if($count == 38) {
-#    exit(0);
-#    }
-     
 }
